% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpd.R
\name{fpd}
\alias{fpd}
\title{calculate the Fritz & Purvis D (FPD) metric}
\usage{
fpd(state, phy)
}
\arguments{
\item{state}{a vector of '0' and '1' for trait state of each tip, in the same order as the tip labels.}

\item{phy}{an object of class 'phylo'.}
}
\description{
This function calculates the FPD metric. To conduct significance test on FPD, use 'fpd' as input of 'func' in 'treestat'
}

