% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noto.R
\name{noto}
\alias{noto}
\title{calculate the number of tips per origin (NoTO) metric}
\usage{
noto(state, phy)
}
\arguments{
\item{state}{a vector of '0' and '1' for trait state of each tip, in the same order as the tip labels.}

\item{phy}{an object of class 'phylo'.}
}
\description{
This function calculates the NoTO metric. To conduct significance test on NoTO, use 'noto' as input of 'func' in 'treestat'
}

