% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscd.R
\name{sscd}
\alias{sscd}
\title{calculate the sum of sister clade differences (SSCD) metric}
\usage{
sscd(state, phy)
}
\arguments{
\item{state}{a vector of '0' and '1' for trait state of each tip, in the same order as the tip labels.}

\item{phy}{an object of class 'phylo'.}
}
\description{
This function calculates the SSCD metric. To conduct significance test on SSCD, use 'sscd' as input of 'func' in 'treestat'
}

