% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tars.R
\name{tars}
\alias{tars}
\title{calculate the tip age rank sum (TARS) metric}
\usage{
tars(state, phy)
}
\arguments{
\item{state}{a vector of '0' and '1' for trait state of each tip, in the same order as the tip labels.}

\item{phy}{an object of class 'phylo'.}
}
\description{
This function calculates the TARS metric. To conduct significance test on TARS, use 'tars' as input of 'func' in 'treestat'
}

