% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/fpd.R
\name{fpd}
\alias{fpd}
\title{calculate the Fritz & Purvis D (FPD) metric}
\usage{
fpd(state, phy)
}
\description{
This function calculates the FPD metric. To conduct significance test on FPD, use 'fpd' as input of 'func' in 'treestat'
}

