% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/noto.R
\name{noto}
\alias{noto}
\title{calculate the number of tips per origin (NoTO) metric}
\usage{
noto(state, phy)
}
\description{
This function calculates the NoTO metric. To conduct significance test on NoTO, use 'noto' as input of 'func' in 'treestat'
}

