% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sscd.R
\name{sscd}
\alias{sscd}
\title{calculate the sum of sister clade differences (SSCD) metric}
\usage{
sscd(state, phy)
}
\description{
This function calculates the SSCD metric. To conduct significance test on SSCD, use 'sscd' as input of 'func' in 'treestat'
}

