% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/tars.R
\name{tars}
\alias{tars}
\title{calculate the tip age rank sum (TARS) metric}
\usage{
tars(state, phy)
}
\description{
This function calculates the TARS metric. To conduct significance test on TARS, use 'tars' as input of 'func' in 'treestat'
}

